Sorting dates when date is inside a link
The issue is seen on http://jsfiddle.net/dnsL2oc4/1/ (with the current `datetime-moment.js`): it is impossible to properly sort the dates since they're inside a link. Instead, the link (`a href`) is used to perform sorting.
The solution can be found on http://jsfiddle.net/dnsL2oc4/2/ and is reflected in this pull request.

See http://stackoverflow.com/questions/28864853/ for background.