Fix broken update on changing input box value
The table would not update if you changed the page value in the input box, i.e. if the input box showed page 1 and you changed it to page 5, nothing would happen whereas you would expect page 5 to be loaded into your table.

`fnExtStylePagingInfo` is used in 3 other places in this file, so I guessed that the usage of `fnPagingInfo` in the `nPageNumBox.change` block was a typo and should have been `fnExtStylePagingInfo`. This indeed fixes the problem.