Fix datetime-luxon.js
luxon.fromFormat argument text must really be text so if the d variable
is not do not call luxon.fromFormat as then we will get the dreaded
TypeError: input.match is not a function. When the variable is not text
it will not match anyway so return null in that case.