Input paging: Styling support for the libraries DataTables works with
I don't know if this is a good way of doing it or not. It does confuse
the code somewhat, but it keeps it all as a single file. The paging
markup is different in each of the styling frameworks, and this isn't
perfect for some - Foundation for example wants the elements changed
when there is an item disabled. DataTables does it by redrawing each
time, but that isn't an option here.

One possible alternative would be to split it to have two seaprate
paging controls (before and after the input). Might be that would be
better?